리덕스툴킷 적용
