+++ package.json
@@ -48,6 +48,7 @@
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@jamesives/github-pages-deploy-action": "^4.2.3",
+    "@reduxjs/toolkit": "^1.8.1",
     "bootstrap": "^5.1.3",
     "facepaint": "^1.2.1",
     "jest-plugin-context": "^2.7.0",
