+++ src/containers/situationSelect/SituationSelectContainer.jsx
@@ -6,7 +6,7 @@ import { useDispatch } from 'react-redux';
 
 import {
   setSituationFilter,
-} from '../../actions';
+} from '../../slice';
 
 const mq = facepaint([
   '@media (min-width: 1024px)',
