+++ src/pages/CustomPage.jsx
@@ -18,7 +18,7 @@ import TopBar from '../components/topbar/TopBar';
 
 import {
   setRestaurantsData,
-} from '../actions';
+} from '../slice';
 
 const mq = facepaint([
   '@media (min-width: 1024px)',
