+++ src/pages/SearchResultRestaurantPage.jsx
@@ -19,7 +19,7 @@ import { saveItem, loadItem } from '../services/storage'
 
 import {
   loadResultRestaurants,
-} from '../actions';
+} from '../slice';
 
 const mq = facepaint([
   '@media (min-width: 1024px)',
